llm-server/pyproject.toml
[tool.poetry]
name = "llm-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.2.60"
fastapi = "^0.115.6"
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
langchain-milvus = "^0.1.7"
langfuse = "^2.57.1"
pymongo = "^4.10.1"
praw = "^7.8.1"
streamlit = "^1.41.1"
opendartreader = "^0.2.3"
pymilvus = "^2.5.1"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.3.1"
uvicorn = "^0.34.0"
huggingface-hub = "^0.27.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
